fix: refactor uploadEncryptedFile sample into its own file
